Add smoke effect when changing player skin,lil scripts fixes
Conditions checks relocated from MonoBehaviour Update method to events

+++ Coffee Addiction/Assets/Scripts/CoffeePickUpController.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class CoffeePickUpController : MonoBehaviour
 {
     private bool isPickUpAllowed;
-    private const float HealingAmount = 20;
+    private const int HealingAmount = 20;
     [SerializeField] private PlayerHealth playerHealthScript;
 
     private void Update()
